Finished the Signup/Login System.
You can now create an account and login.

Added alot of valdiation for signup & login.

User data will now be stored in a database.

Added a PageManager navigates you to a new page and
brings you back to a pervious page by storing certain
pages onto a stack. NOTE: ONLY PAGES THAT YOU NEED TO
RETURN TOO GET STORED ON THE STACK!

Was meant to do small pushes but got carried away
and decided to focus on the entire thing lol.

Going to focus on creating the proper database now
and seperate the queries into a properties file.